+++ ving/docs/change-log.md
@@ -5,6 +5,9 @@ outline: deep
 
 ## July 2024
 
+### 2024-07-04
+* Added `options` param to ving schemas.
+
 ### 2024-07-01
 * Added verifiedEmail field to the User CLI.
 * Fixed a problem with creating new users via the rest API.